add Native auth support, move req into new file
