+++ go.mod
@@ -3,3 +3,5 @@ module golift.io/starr
 go 1.17
 
 require github.com/golang/mock v1.6.0
+
+require golang.org/x/net v0.0.0-20211216030914-fe4d6282115f // indirect
